+++ lib/codeowners_rule.ex
@@ -3,8 +3,29 @@ defmodule Codeowners.Rule do
   Contains types and functions related to individual CODEOWNERS rules.
   """
 
+  alias Codeowners.Rule
+
   @type t :: %Codeowners.Rule{pattern: String.t(), regex: Regex.t(), owners: list(String.t())}
   defstruct pattern: nil, regex: nil, owners: []
+
+  @doc """
+  Build a `Codeowners.Rule` for the given CODEOWNERS line.
+  """
+  @spec regex(String.t()) :: t() | nil
+  def build(line) when is_binary(line) do
+    rule =
+      String.split(line, "#")
+      |> List.first()
+
+    case rule do
+      "" ->
+        nil
+
+      _ ->
+        [pattern | owners] = String.split(rule)
+        %Rule{pattern: pattern, regex: Rule.regex(pattern), owners: owners}
+    end
+  end
 
   @doc """
   Compile a `Regex` for the given pattern string.